+++ seed/template/angularjs/app/components/table/line/box/box.ts
@@ -11,6 +11,8 @@ angular
 
     controller() {
         this.push = () => {
+            if (this.xo !== undefined) return
+
             this.ssTable.xo = (this.ssTable.xo === 'O') ? 'X' : 'O'
 
             this.xo = this.ssTable.xo